move fixture into subfolder
