+++ fixtures/fixture.html
@@ -5,7 +5,7 @@
 		<title>Fixture</title>
 		<script>
 			window.addEventListener('load', () => {
-				const contextMenu = require('.');
+				const contextMenu = require('..');
 
 				contextMenu({
 					window: document.getElementById('webview')
